kyc_ocr_extraction/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ocr-pro==0.1.3
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pymongo==4.15.5
pytesseract==0.3.13
Werkzeug==3.1.4
